Feat :sparkles: Implemetação de recursos para gerenciamento de empresas.

+++ src/modules/products/products.service.ts
@@ -6,33 +6,45 @@ import { Repository } from 'typeorm';
 import { UpdateProductDto } from './dto/update-product.dto';
 import { ProductsQueryPaginateDto } from './dto/products-query-paginate.dto';
 import { ProductsOutputPaginateDto } from './dto/products-output-paginate.dto';
+import { CompaniesService } from '../companies/companies.service';
 
 @Injectable()
 export class ProductsService {
     constructor(
         @Inject(PRODUCT_REPOSITORY)
         private productRepository: Repository<Product>,
+        private readonly companiesService: CompaniesService,
     ) {}
 
-    async create(createProductDto: CreateProductDto): Promise<Product> {
-        const product = this.productRepository.create(createProductDto);
+    async create(companyId: string, createProductDto: CreateProductDto): Promise<Product> {
+        const company = await this.companiesService.findOne(+companyId);
+        const product = this.productRepository.create({ ...createProductDto, company });
         return await this.productRepository.save(product);
     }
 
-    async findOne(id: number): Promise<Product> {
-        const product = await this.productRepository.findOne({ where: { id } });
+    async findOne(companyId: string, id: number): Promise<Product> {
+        await this.companiesService.findOne(+companyId);
+        const product = await this.productRepository.findOne({
+            where: {
+                id,
+                company: { id: +companyId },
+            },
+        });
         if (!product) {
-            throw new NotFoundException(`Product with ID ${id} not found`);
+            throw new NotFoundException(`Product with ID ${id} not found in Company ID ${companyId}`);
         }
         return product;
     }
 
-    async paginate(queryDto: ProductsQueryPaginateDto): Promise<ProductsOutputPaginateDto> {
+    async paginate(companyId: string, queryDto: ProductsQueryPaginateDto): Promise<ProductsOutputPaginateDto> {
         const { pagination, page, limit, search } = queryDto;
-        const queryBuilder = this.productRepository.createQueryBuilder('product');
+
+        await this.companiesService.findOne(+companyId);
+
+        const queryBuilder = this.productRepository.createQueryBuilder('product').where('product.companyId = :companyId', { companyId: +companyId }); // Filtro pela empresa
 
         if (search) {
-            queryBuilder.where('product.name LIKE :search OR product.description LIKE :search', { search: `%${search}%` });
+            queryBuilder.andWhere('product.name LIKE :search OR product.description LIKE :search', { search: `%${search}%` });
         }
 
         if (pagination) {
@@ -56,22 +68,30 @@ export class ProductsService {
         };
     }
 
-    async update(id: number, updateProductDto: UpdateProductDto): Promise<Product> {
-        const updateResult = await this.productRepository.update(id, updateProductDto);
-        if (updateResult.affected === 0) {
-            throw new NotFoundException(`Product with ID ${id} not found`);
+    async update(companyId: string, id: number, updateProductDto: UpdateProductDto): Promise<Product> {
+        await this.companiesService.findOne(+companyId);
+
+        const product = await this.productRepository.findOne({ where: { id, company: { id: +companyId } } });
+        if (!product) {
+            throw new NotFoundException(`Product with ID ${id} not found in Company ID ${companyId}`);
         }
-        const updatedProduct = await this.productRepository.findOne({ where: { id } });
+
+        await this.productRepository.update(id, updateProductDto);
+
+        const updatedProduct = await this.productRepository.findOne({ where: { id, company: { id: +companyId } } });
         if (!updatedProduct) {
-            throw new NotFoundException(`Updated product with ID ${id} not found`);
+            throw new NotFoundException(`Updated product with ID ${id} not found in Company ID ${companyId}`);
         }
+
         return updatedProduct;
     }
 
-    async remove(id: number): Promise<void> {
-        const deleteResult = await this.productRepository.delete(id);
-        if (deleteResult.affected === 0) {
-            throw new NotFoundException(`Product with ID ${id} not found`);
+    async remove(companyId: string, id: number): Promise<void> {
+        await this.companiesService.findOne(+companyId);
+        const product = await this.productRepository.findOne({ where: { id, company: { id: +companyId } } });
+        if (!product) {
+            throw new NotFoundException(`Product with ID ${id} not found in Company ID ${companyId}`);
         }
+        await this.productRepository.delete(id);
     }
 }
